+++ App.tsx
@@ -4,44 +4,15 @@ import { SafeAreaProvider } from 'react-native-safe-area-context';
 import useCachedResources from './hooks/useCachedResources';
 import useColorScheme from './hooks/useColorScheme';
 import Navigation from './navigation';
-import * as Notifications from 'expo-notifications';
 
-//import { registerDevicePushTokenAsync } from '../api';
-import Constants from 'expo-constants';
-import { Linking, Platform } from 'react-native';
-import { StorePushToken } from './storage/pushNotificationStorage';
-
-import { View, StyleSheet } from 'react-native';
 
 export default function App() {
   const isLoadingComplete = useCachedResources();
   const colorScheme = useColorScheme();
-  const lastNotificationResponse:any = Notifications.useLastNotificationResponse();
-
-
+ 
   if (!isLoadingComplete) {
     return null;
   } else {
-
-
-  // useEffect(() => {
-  //   if (
-  //     lastNotificationResponse &&
-  //     lastNotificationResponse.notification.request.content.data.url &&
-  //     lastNotificationResponse.actionIdentifier === Notifications.DEFAULT_ACTION_IDENTIFIER
-  //   ) {
-  //     Linking.openURL(lastNotificationResponse.notification.request.content.data.url);
-  //   }
-  // }, [lastNotificationResponse]);
-
-  // useEffect(() => {
-  //   registerForPushNotificationsAsync();
-  //   //registerForPushNotificationsAsync().then(token => StorePushToken(token as string));
-  //   schedulePushNotification();
-
-  // }, []);
-
-
 
     return (
       <SafeAreaProvider>
@@ -52,45 +23,3 @@ export default function App() {
   }
 }
 
-async function schedulePushNotification() {
-  const identifyer = await Notifications.scheduleNotificationAsync({
-    content: {
-      title: "🌻🌻🌻🌻🌻",
-      body: 'A message of positivity',
-      data: { url: 'https://miricyl.org' },
-    },
-    trigger: { seconds: 2 },
-  });
-  console.log(identifyer);
-}
-
-async function registerForPushNotificationsAsync() {
-  let token;
-  if (Constants.isDevice) {
-    const { status: existingStatus } = await Notifications.getPermissionsAsync();
-    let finalStatus = existingStatus;
-    if (existingStatus !== 'granted') {
-      const { status } = await Notifications.requestPermissionsAsync();
-      finalStatus = status;
-    }
-    if (finalStatus !== 'granted') {
-      alert('Failed to get push token for push notification!');
-      return;
-    }
-    token = (await Notifications.getExpoPushTokenAsync()).data;
-    console.log(token);
-  } else {
-    alert('Must use physical device for Push Notifications');
-  }
-
-  if (Platform.OS === 'android') {
-    Notifications.setNotificationChannelAsync('default', {
-      name: 'default',
-      importance: Notifications.AndroidImportance.MAX,
-      vibrationPattern: [0, 250, 250, 250],
-      lightColor: '#FF231F7C',
-    });
-  }
-
-  return token;
-}
