+++ screens/DashboardScreen.tsx
@@ -1,12 +1,35 @@
+import Constants from 'expo-constants';
 import * as React from 'react';
-import { StyleSheet, ImageBackground, Image } from 'react-native';
+import { useEffect } from 'react';
+import { StyleSheet, ImageBackground, Image, Linking, Platform } from 'react-native';
 import { ScrollView } from 'react-native-gesture-handler';
 import NavigationCard from '../components/NavigationCard';
 import { Text, View } from '../components/Themed';
 import Layout from '../constants/Layout';
 import {LinkType} from '../types'
+import { StorePushToken } from './storage/pushNotificationStorage';
+import * as Notifications from 'expo-notifications'
 
 const DashboardScreen = () => {
+
+  //const lastNotificationResponse:any = Notifications.useLastNotificationResponse();
+
+  //   useEffect(() => {
+  //   if (
+  //     lastNotificationResponse &&
+  //     lastNotificationResponse.notification.request.content.data.url &&
+  //     lastNotificationResponse.actionIdentifier === Notifications.DEFAULT_ACTION_IDENTIFIER
+  //   ) {
+  //     Linking.openURL(lastNotificationResponse.notification.request.content.data.url);
+  //   }
+  // }, [lastNotificationResponse]);
+
+  useEffect(() => {
+    registerForPushNotificationsAsync();
+    //registerForPushNotificationsAsync().then(token => StorePushToken(token as string));
+    schedulePushNotification();
+
+  }, []);
   
   return (
       <View style={styles.screen}>
@@ -68,3 +91,47 @@ const styles = StyleSheet.create({
     width: Layout.window.width * 0.30,
   }
 });
+
+
+async function schedulePushNotification() {
+  const identifyer = await Notifications.scheduleNotificationAsync({
+    content: {
+      title: "🌻🌻🌻🌻🌻",
+      body: 'A message of positivity',
+      data: { url: 'https://miricyl.org' },
+    },
+    trigger: { seconds: 2 },
+  });
+  console.log(identifyer);
+}
+
+async function registerForPushNotificationsAsync() {
+  let token;
+  if (Constants.isDevice) {
+    const { status: existingStatus } = await Notifications.getPermissionsAsync();
+    let finalStatus = existingStatus;
+    if (existingStatus !== 'granted') {
+      const { status } = await Notifications.requestPermissionsAsync();
+      finalStatus = status;
+    }
+    if (finalStatus !== 'granted') {
+      alert('Failed to get push token for push notification!');
+      return;
+    }
+    token = (await Notifications.getExpoPushTokenAsync()).data;
+    console.log(token);
+  } else {
+    alert('Must use physical device for Push Notifications');
+  }
+
+  if (Platform.OS === 'android') {
+    Notifications.setNotificationChannelAsync('default', {
+      name: 'default',
+      importance: Notifications.AndroidImportance.MAX,
+      vibrationPattern: [0, 250, 250, 250],
+      lightColor: '#FF231F7C',
+    });
+  }
+
+  return token;
+}
